Add -P <pid-file> option to ns_exec
ns_exec knows the pid of the container-init in its own namespace (this
is usually the global pid since we normally run ns_exec in init-pid-ns).
If ns_exec writes out this "global" pid to a file, wrapper scripts can
find the process they created more easily.

killall and pidof assume that all previous instances of the test program
have thawed and exited gracefully :-) As such they are not very friendly
during early testing and/or complicate the wrapper scripts.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>