Add test case ptree1 - process tree using fileio
The test case continously copies a given input-file to an output-file,
one record at a time. After copying, the test verifies that the two files
have the same data (i.e ensures if test was checkpointed/restarted, the
copy was still correct).  The test ends when the 'test_done()' condition
is TRUE.

The accompanying wrapper runs the test in a loop and checkpoints/restarts
the test.

Changelog:
	(based on review comments from Serge Hallyn)
	- change checkpoint_ready() to set_checkpoint_ready()
	- allocate memory dynamically for the data and log file names
	  (since the names depend on depth).
	- Move data files to 'data.d' directory

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>