libcrtest: define and use freezer helpers
1. rename common.h to libcrtest.h
2. make a libcrtest.a library out of common.o
3. define logfp in common.c and no tin ptree1.c
4. define helper to move a task to a freezer cgroup
5. use helper from (4) in sleeptest.c and userns/*.c

This removes the need for freezer to be mounted at the hardcoded
location '/cgroup'.

I suppose the library could also go ahead and mount the freezer
if not already done?  I will wait for someone to ask for that
before doing it, as I don't like my tests being too smart and
wresting away control of my systems!

Signed-off-by: Serge Hallyn <serue@us.ibm.com>