c/r tests: Add futex c/r tests
	Add tests for plain, robust, and pi futexes. Each test sets up a
typical contended futex scenario and then awaits checkpoint. We only test
the contended case since the uncontended cases are entirely based on the
state of userspace memory. After checkpoint each test verifies that the
critical semantics of the futex still works.

	For plain futexes we ensure that the same number of tasks that
were asleep on the futex are woken up.

	For robust futexes we set the robust list head of each process
and wait for checkpoint. After checkpoint we verify that the kernel
still knows about the robust list head then each child exits without
releasing the futex. Since the child still holds the futex at exit the
kernel wakes another waiting child.

	For pi futexes we set up the contended priority inversion case
which is supposed to cause priority inheritance. Then we wait for checkpoint.
After checkpoint we verify that the priority is inherited and we also
check that the remaining waiters are woken in priority order.

	We do not test some variations on these such as private futexes,
bitsets, requeing, and futexes mapped in filesystem files. All of the futexes
in these tests are in anonymous shared mappings.

	README.txt describes the tests and their log output. run.sh shows
how to run these tests.

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>