futex: make 'plain' handle -EINTR properly (v2)
If one of 'plain's' kids gets a signal (or gets -EINTR
for some other reason, with freeze; ckpt; thaw) then
futex_wait may restart after the parent has set futex->counter
to 1.  In that case, futex_wait will return -EAGAIN, and kid
is coded such that it will keep retrying futex_wait in an
endless loop.

Fix that.  If futex_wait returned -EINVAL, and if futex->counter
is 1, then assume we're supposed to exit.

Also, in that case the parent's futex_wait will return some
number less than number of kids.  But futex_wait return value
is an error only if it's -1.  So accept a return value less
than N.

Changelog:
	Aug 3: As pointed out by Matt Helsley, decrement i
	       if malloc or clone fails, to avoid a bad kill
	       on cleanup.

Signed-off-by: Serge Hallyn <serue@us.ibm.com>