cr_tests: Add timeout to robust futex test.
The robust futex test can hang if the kernel fails to properly set the
robust list pointer. This currently happens during restart. The test should not
hang and instead should report failure.

Use a timeout to ensure that hangs are caught and reported as failure.
The timeout should return ETIMEDOUT. This limits the total amount of
time checkpoint/restart can take so a suitable timeout is essential here.

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>
Cc: containers@lists.linux-foundation.org