add a simple credentials restore check, and a bugfix
Two ipc checks added cap_sys_admin+pe to restart, but did not
remove it on error path.

A simple test checkpointing nested user namespaces was there but
not being executed.  Execute it.

Also add a test which creates a checkpoint for uid 0, then tries
to restart it as the ltp user.

Signed-off-by: Serge E. Hallyn <serue@us.ibm.com>