cr_tests: epoll: Add two testcases for epoll checkpoint/restart.
The first, "empty" creates an empty epoll set, waits for checkpoint,
and closes it afterwards.

The second test creates and epoll set, adds the write end of a pipe,
waits for EPOLLOUT (writability) with epoll, writes, adds the
read end, removes the write end, waits for EPOLLIN (readability),
reads, and closes the epoll set.

Finally the patch adds a run.sh script much like the one for futexes.
This one contains modifications which run all of the "subtests" of
the pipe test as well.

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>