cr_tests: epoll: Add sk10k - poll io on ten thousand sockets with epoll
This test opens a specified number of socket pairs, sends a message over
each, reads the message, and exits. It uses epoll to "wait" for events
on the sockets. This tests epoll sets that may stress the watch
limit, the open file descriptor limit, and the limits of epoll c/r
itself.

Since an epoll item c/r struct takes 20 bytes on i386 we need at least
205 sockets to have multiple 4k pages worth of epoll items to
checkpoint.

The goal is to foil the kmalloc used to checkpoint epoll items in the kernel
such that we can't checkpoint them all in one go and without needing to
reproduce OOM conditions. Since we can't rely on PAGE_SIZE to calculate the
number of sockets needed to do this and since we're limited by rlimit
anyhow, we just go with the hard limit for now.

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>