cr_tests: Add eventfd01 test based on LTP test
Add checkpoint spots to the LTP test. Notably different from the
epoll test library, these labels to not incorporate error handling
code. Rather the test and branch are explicit and rely on LTP tst_*
functions. As such it requires libltp.a to link. Do this with:

make LTPROOT=/path/to/ltp/install

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>

[RFC] Due to weirdness/bugs in test results after adapting testcase
	to c/r.