test chroot and chdir restoration
One testcase ensures that cwd is restore and chroot is not
wrongly set.  The second ensure that a container with a
child task which is chrooted, when restored, is restored
with the child task chrooted.

The third testcase tries to restore a container which was
itself chrooted.  This fails, as it is up to userspace to
make sure that the container is restarted in the right
place.  One way to do this is with the
	restart -r chroot_dir < ckpt
option.

(The third testcase does not by default run)

Signed-off-by: Serge Hallyn <serue@us.ibm.com>