filelease2: Test C/R during lease-break-interval
When a process, P1 (lease-owner of a file) is notified of an impending
lease-break (due to a conflicting open of the file by process P2 (lease-
breaker), have P1 "flush" some data to the file. Checkpoint processes
P1 and P2 before the kernel forcibly revokes the lease (i.e checkpoint
within the lease-break-interval).

Upon restart, verify that the P2's open() of the file completes properly
and that the data flushed by P1 is visible to P2.

NOTE:
	This test passes even with out support for checkpoint of file-locks.
	The reason for this is that the kernel terminates the lease of P1
	before notifying P1 about the lease-break. Since we checkpoint while
	P1 is in signal handler, neither P1 nor P2 have a lease and so the
	C/R passes. Hopefully the data-comparison checks will catch some
	errors during development. If not, we may need to redesign the test
	or drop it later.

Signed-off-by: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>