cr_tests: rework build system
Currently we have a diverse set of independent Makefiles with their
own conventions and quirks which are invoked recursively by the
top-level Makefile. This makes it difficult to modify build parameters
that users may expect to change on the command line (e.g. CC or
CFLAGS).  The current state of affairs makes it very difficult to
build 64-bit binaries where the default toolchain output is 32-bit or
vice-versa.  This seriously hinders work on powerpc where we need to
test both 32- and 64-bit programs.

With the system provided by this patch, the user can alter the build's
behavior easily and consistently.  For example, "make CFLAGS=-m64"
just works.  New tests need only provide a module.mk file and set a
few variables and the occasional custom rule.  There are two main behavior
changes:

- "make clean" no longer removes test outputs; a new "testclean"
  target is added for that purpose.  I did this because I know these
  tests are typically run as root and I wanted to minimize the
  possibility of running rm -rf of some vital system directory.

- All tests are built with -Wall, enforced by the top-level Makefile.

Although no regressions are intended, the old Makefiles remain in
place, unmodified by this patch.  To revert to the old system, use
"make -f Makefile".

Signed-off-by: Nathan Lynch <ntl@pobox.com>