cr_tests: futex: pi: Adjust RT priority range
Merely checking the priority values does not work on some systems. For example,
my Fedora 12 system shows default realtime priority limits of 0 (soft) 0 (hard)
even for root.

Instead, read the limits and attempt to set them such that they will work
for the test. If we fail to set them then fail the test as well.

Signed-off-by: Matt Helsley <matthltc@us.ibm.com>
Signed-off-by: Serge Hallyn <serue@us.ibm.com>